References #2302
I've given this out a few times with no complaints. Yes all this patch does
is not use std::vector<int>.

It is time to give up on it being fixed where it should be, with luck
this can be reverted some day.


git-svn-id: http://svn.mythtv.org/svn/trunk@12842 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mythvideo/mythvideo/dbaccess.h
@@ -34,7 +34,7 @@ class MultiValue
     struct entry
     {
         int id;
-        typedef std::vector<int> values_type;
+        typedef std::vector<long> values_type;
         values_type values;
     };
     typedef std::vector<entry> entry_list;
